agent/requirements.txt
openai
python-dotenv
google-search-results
google-generativeai
